move codegen parts of alist.cpp to codegen/
